% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_barcodes.R
\name{correct_barcodes}
\alias{correct_barcodes}
\title{Make barcode whitelist}
\usage{
correct_barcodes(
  whitelist,
  r1_fastq,
  write_dir = ".",
  max_linker_dels = 5,
  batch_size = 10^5
)
}
\arguments{
\item{whitelist}{path to a bead dedupliction map, as created by make_barcode_whitelist()}

\item{r1_fastq}{path to a read 1 fastq file -- may be comma delimited list of fastqs from same puck/library}

\item{write_dir}{(optional) directory to write to, defaults to current directory}

\item{max_linker_dels}{(optional) integer for max acceptable deletions in linker, defaults to 5}

\item{batch_size}{(optional) integer determining number of reads per batch, defaults to 10^5}

\item{n_cores}{(optional) integer indicating number of cores for parallelization, defaults to 1}
}
\value{
none
}
\description{
Use bead deduplication map to generate a whitelist of all acceptable + non-ambiguous barcode matches.
}
\examples{
make_barcode_whtielist(dedup_map='coordinates.tsv_deduplication_map.txt')
make_barcode_whtielist(dedup_map='coordinates.tsv_deduplication_map.txt',write_dir='~/mydir/',n_cores=12)
}
