% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barcode_whitelist.R
\name{make_barcode_whitelist}
\alias{make_barcode_whitelist}
\title{Make barcode whitelist}
\usage{
make_barcode_whitelist(dedup_map, coords_file, write_dir = ".", n_cores = 1)
}
\arguments{
\item{dedup_map}{path to a bead dedupliction map, as created by make_bead_dedup_map()}

\item{coords_file}{path to a tab-delimited file containing the barcodes and coordinates for the puck}

\item{write_dir}{(optional) directory to write to, defaults to current directory}

\item{n_cores}{(optional) integer indicating number of cores for parallelization, defaults to 1}
}
\value{
none
}
\description{
Use bead deduplication map to generate a whitelist of all acceptable + non-ambiguous barcode matches.
}
\examples{
make_barcode_whtielist(dedup_map='coordinates.tsv_deduplication_map.txt')
make_barcode_whtielist(dedup_map='coordinates.tsv_deduplication_map.txt',write_dir='~/mydir/',n_cores=12)
}
