% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bead_dedup_map.R
\name{make_bead_dedup_map}
\alias{make_bead_dedup_map}
\title{Make bead deduplication map}
\usage{
make_bead_dedup_map(coords_file, write_dir = ".", max_slide_dist = 10)
}
\arguments{
\item{coords_file}{path to a tab-delimited file containing the barcodes and coordinates for the puck}

\item{write_dir}{(optional) directory to write to, defaults to current directory}

\item{max_slide_dist}{(optional) maximum slide distance to be considered a duplication, defaults to 10}
}
\value{
none
}
\description{
Identify groups of virtually duplicted beads and re-direct each group to a single barcode, write out text file.
}
\examples{
make_bead_dedup_map(coords_file='coordinates.tsv')
make_bead_dedup_map(coords_file='coordinates.tsv',write_dir='~/mydir/',max_linker_dels=3)
}
