% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syrah.R
\name{syrah}
\alias{syrah}
\title{Syrah}
\usage{
syrah(
  coords_file,
  r1_fastq,
  write_dir = ".",
  n_cores = 1,
  max_slide_dist = 1,
  max_linker_dels = 5,
  batch_size = 10^5
)
}
\arguments{
\item{coords_file}{path to a tab-delimited file containing the barcodes and coordinates for the puck}

\item{r1_fastq}{path to a read 1 fastq file -- may be comma delimited list of fastqs from same puck/library}

\item{write_dir}{(optional) directory to write to, defaults to current directory}

\item{n_cores}{(optional) integer indicating number of cores for parallelization, defaults to 1}

\item{max_slide_dist}{(optional) maximum slide distance to be considered a duplication, defaults to 10}

\item{max_linker_dels}{(optional) integer for max acceptable deletions in linker, defaults to 5}

\item{batch_size}{(optional) integer determining number of reads per batch, defaults to 10^5}
}
\value{
none
}
\description{
Wrapper for make_bead_dedup_map() + make_barcode_whitelist() + correct_barcodes()
}
\examples{
syrah('coordinates.txt','read1.fastq')
syrah(coords_file='coordinates.txt',r1_fastq='read1.fastq,other_read1.fastq',write_dir='~/mydir/',n_cores=12)
}
